Extracted test_copy_form
This one isn't even related to suites.
